# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgniAI API",
    "description": "REST API for AgniAI — Offline Agniveer Chatbot with RAG pipeline and Admin intelligence layer.\n\n## User APIs\nChat with the Agniveer knowledge base, ingest documents, manage sessions.\n\n## Admin APIs\nNatural-language queries that route through the .NET AiCommand backend for structured data (Performance, Leave, Medical, Attendance, Verification, Equipment, Distribution, Skills).",
    "version": "1.0.0",
    "contact": {
      "name": "AgniAI"
    }
  },
  "servers": [],
...
  "paths": {
...
    "/api/chat": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Chat (RAG)",
        "description": "Send a message to the AgniAI chatbot. Automatically classifies intent (chat / RAG / reject), retrieves relevant chunks from FAISS+BM25, and generates a grounded answer via Ollama.\n\n**Streaming:** Set `stream: true` and include header `Accept: text/event-stream` for SSE token streaming.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Answer generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimit"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
    "/api/admin/classify": {
      "post": {
        "tags": [
          "Admin Chat"
        ],
        "summary": "Classify admin intent (debug)",
        "description": "Runs the intent classifier and returns the structured JSON payload — **without** calling the .NET backend. Use this to debug how a question is being interpreted before wiring up .NET.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminClassifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Intent classification result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminClassifyResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/admin/chat": {
      "post": {
        "tags": [
          "Admin Chat"
        ],
        "summary": "Admin chat",
        "description": "Full admin pipeline:\n1. Classify intent → structured JSON\n2. POST to `.NET /api/AiCommand/execute`\n3. Format response → human-readable answer\n\n**Supported modules:** Performance, Leave, Medical, Attendance, Verification, Equipment, Distribution, Skills/Roster",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Formatted answer from .NET data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminChatResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimit"
          },
          "502": {
            "description": "Bad gateway — .NET backend error or unreachable",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Required only for /api/reset_index when API_SECRET_KEY is set in .env"
      }
    },
    "schemas": {
      "ChatRequest": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "User question"
          },
          "model": {
            "type": "string",
            "description": "Ollama model override"
          },
          "stream": {
            "type": "boolean",
            "description": "Enable SSE streaming (requires Accept: text/event-stream header)",
            "default": false
          },
          "session_id": {
            "type": "string",
            "description": "Session identifier for conversation memory"
          }
        }
      },
      "ChatResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "answer": {
            "type": "string",
            "description": "The generated answer"
          },
          "style": {
            "type": "string",
            "enum": [
              "short",
              "elaborate",
              "detail"
            ],
            "description": "Answer style detected from the query"
          },
          "session_id": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "AdminChatRequest": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          }
        }
      },
      "AdminChatResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean"
          },
          "introMessage": {
            "type": "object"
          },
          "formattedData": {
            "type": "object"
          },
          "analysis": {
            "type": "object"
          },
          "prediction": {
            "type": "object"
          },
          "conclusion": {
            "type": "object"
          },
          "suggestedQuestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "widgets": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "metadata": {
            "type": "object"
          },
          "overallConfidence": {
            "type": "number"
          },
          "partialFailure": {
            "type": "boolean"
          },
          "failedSections": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AdminClassifyRequest": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "AdminClassifyResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "intent": {
            "$ref": "#/components/schemas/FrontendIntent"
          },
          "dotnet_payload": {
            "$ref": "#/components/schemas/DotnetPayload"
          }
        }
      },
...
      "FrontendIntent": {
        "type": "object",
        "description": "Frontend-facing intent view (request fields plus visual type)",
        "properties": {
          "commandId": {
            "type": "integer",
            "default": 0
          },
          "category": {
            "type": "string"
          },
          "operation": {
            "type": "string"
          },
          "section": {
            "type": "string"
          },
          "subSection": {
            "type": "string"
          },
          "grading": {
            "type": "string"
          },
          "attemptNo": {
            "type": "integer"
          },
          "fromAttempt": {
            "type": "integer"
          },
          "toAttempt": {
            "type": "integer"
          },
          "leaveType": {
            "type": "string"
          },
          "bmiCategory": {
            "type": "string"
          },
          "bloodGroup": {
            "type": "string"
          },
          "equipmentName": {
            "type": "string"
          },
          "n": {
            "type": "integer"
          },
          "batchId": {
            "type": "integer"
          },
          "platoonId": {
            "type": "integer"
          },
          "unitName": {
            "type": "string"
          },
          "agniveerNo": {
            "type": "string"
          },
          "companyId": {
            "type": "integer"
          },
          "sport": {
            "type": "string"
          },
          "class": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "fromDate": {
            "type": "string",
            "format": "date-time"
          },
          "toDate": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string",
            "description": "Frontend widget type"
          }
        }
      },
      "DotnetPayload": {
        "type": "object",
        "description": "Payload sent to .NET /api/AiCommand/execute (null fields stripped)",
        "properties": {
          "commandId": {
            "type": "integer",
            "default": 0
          },
          "category": {
            "type": "string"
          },
          "operation": {
            "type": "string"
          },
          "section": {
            "type": "string"
          },
          "subSection": {
            "type": "string"
          },
          "grading": {
            "type": "string"
          },
          "attemptNo": {
            "type": "integer"
          },
          "fromAttempt": {
            "type": "integer"
          },
          "toAttempt": {
            "type": "integer"
          },
          "leaveType": {
            "type": "string"
          },
          "bmiCategory": {
            "type": "string"
          },
          "bloodGroup": {
            "type": "string"
          },
          "equipmentName": {
            "type": "string"
          },
          "n": {
            "type": "integer"
          },
          "batchId": {
            "type": "integer"
          },
          "platoonId": {
            "type": "integer"
          },
          "unitName": {
            "type": "string"
          },
          "agniveerNo": {
            "type": "string"
          },
          "companyId": {
            "type": "integer"
          },
          "sport": {
            "type": "string"
          },
          "class": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "fromDate": {
            "type": "string",
            "format": "date-time"
          },
          "toDate": {
            "type": "string",
            "format": "date-time"
          },
          "fullName": {
            "type": "string"
          },
          "groupBy": {
            "type": "string"
          },
          "analyticsHint": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request — missing or invalid fields",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "RateLimit": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "LLM service unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}